refactor(qms): extract ISO17025_TREE data to lib/qms-tree.ts, remove @ts-nocheck
Monday refactor angle — 2026-06-16.

qms/page.tsx was 893 lines and carried @ts-nocheck, with 290 lines of the
ISO 17025:2017 clause-tree data (DocLevel/ClauseNode types, LEVEL_LABELS,
ISO17025_TREE constant, flattenDocs helper, ALL_DOCS, docStatusColors) inlined
directly in the component file.

Extract all of the above to a new lib/qms-tree.ts module:
- DocLevel and ClauseNode interfaces are now exported (ClauseNode.icon typed
  as LucideIcon instead of the previous `any`)
- LEVEL_LABELS typed as Record<1|2|3|4, ...> so indexing by doc.level is
  type-safe
- docStatusColors typed as Record<DocLevel['status'], string> — exhaustive
  key check at compile time
- ISO17025_TREE, flattenDocs, and ALL_DOCS exported for use by the page and
  any future consumers (e.g. a search endpoint or compliance report)

qms/page.tsx now:
- Has no @ts-nocheck — TypeScript checks the full file
- Imports the types/data from @/lib/qms-tree
- Drops ~330 lines; component logic is clearly separated from static data
- Unused lucide imports (ClipboardList, FileCheck, Settings, Users, Search,
  Input, CardDescription) removed from the import list

No behaviour change.
